refactor(solver): split flow comparability queries
